+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.13] - 2024-11-19
+### Added
+- Conditional radios
+### Changed
+- Added `params.hideFormControlError` in elements `checkboxes`, `fileInput`, `radios`, `select`, `textInput` to hide the form control error (red line on the left). Mostly used in conditional radio elements.
+
 ## [v0.0.13] - 2024-11-05
 ### Added
 - Summary list component